+++ src/Fasaani/Config.fs
@@ -1,0 +1,6 @@
+[<AutoOpen>]
+module Fasaani.Config
+
+let defaultConfig (logger : string -> unit) =
+    { Log = Some (QueryDetails.ToString >> logger)
+      CancellationToken = None }